+++ sdk/core/azure_core/src/http/policies/mod.rs
@@ -8,7 +8,7 @@ mod client_request_id;
 mod instrumentation;
 mod user_agent;
 
-pub use bearer_token_policy::BearerTokenAuthorizationPolicy;
+pub use bearer_token_policy::{Authorizer, BearerTokenAuthorizationPolicy, OnChallenge, OnRequest};
 pub use client_request_id::*;
 pub use instrumentation::*;
 pub use typespec_client_core::http::policies::*;